Rearrange package structure
